backend/requirements.txt
fastapi==0.110.1
httpx==0.28.1
motor==3.3.1
pydantic==2.12.5
pymongo==4.5.0
python-dotenv==1.2.2
uvicorn[standard]==0.25.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
